Visual Studio 2008/RestApiSDK/PayPal/Api/Payments/Amount.cs
using System;
using Newtonsoft.Json;
using Newtonsoft.Json.Serialization;
using PayPal.Api.Payments;

namespace PayPal.Api.Payments
{

	/// <summary>
	/// 
    /// </summary>
	public class Amount : Resource  
	{

		/// <summary>
		/// total
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string total
		{
			get;
			set;
		}
		

		/// <summary>
		/// currency
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public string currency
		{
			get;
			set;
		}
		

		/// <summary>
		/// details
    	/// </summary>
		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
		public AmountDetails details
		{
			get;
			set;
		}
		

		/// <summary>
		/// Converts the object to JSON string
		/// </summary>
		public new string ConvertToJson() 
    	{ 
    		return JsonFormatter.ConvertToJson(this);
    	}
    	
	}
}
